requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
altair==6.0.0
    # via streamlit
annotated-doc==0.0.4
    # via typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.14.3
    # via egovault (pyproject.toml)
blinker==1.9.0
    # via streamlit
cachetools==7.0.5
    # via streamlit
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via
    #   pdfminer-six
    #   requests
click==8.3.2
    # via
    #   egovault (pyproject.toml)
    #   streamlit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==46.0.7
    # via
    #   google-auth
    #   pdfminer-six
    #   pyjwt
ebooklib==0.20
    # via egovault (pyproject.toml)
et-xmlfile==2.0.0
    # via openpyxl
filelock==3.25.2
    # via
    #   huggingface-hub
    #   torch
fsspec==2026.3.0
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-api-core==2.30.3
    # via google-api-python-client
google-api-python-client==2.194.0
    # via egovault (pyproject.toml)
google-auth==2.49.2
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.3.1
    # via google-api-python-client
google-auth-oauthlib==1.3.1
    # via egovault (pyproject.toml)
googleapis-common-protos==1.74.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.4.3
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   huggingface-hub
    #   mcp
httpx-sse==0.4.3
    # via mcp
huggingface-hub==1.10.1
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.5.3
    # via scikit-learn
jsonschema==4.26.0
    # via
    #   altair
    #   mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
lxml==6.0.4
    # via
    #   ebooklib
    #   egovault (pyproject.toml)
    #   python-docx
    #   python-pptx
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mcp==1.27.0
    # via egovault (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
narwhals==2.19.0
    # via altair
networkx==3.6.1
    # via torch
numpy==2.4.4
    # via
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   streamlit
    #   transformers
oauthlib==3.3.1
    # via requests-oauthlib
openpyxl==3.1.5
    # via egovault (pyproject.toml)
packaging==26.0
    # via
    #   altair
    #   huggingface-hub
    #   streamlit
    #   transformers
pandas==3.0.2
    # via
    #   egovault (pyproject.toml)
    #   streamlit
pdfminer-six==20251230
    # via pdfplumber
pdfplumber==0.11.9
    # via egovault (pyproject.toml)
pillow==12.2.0
    # via
    #   pdfplumber
    #   python-pptx
    #   streamlit
prompt-toolkit==3.0.52
    # via egovault (pyproject.toml)
proto-plus==1.27.2
    # via google-api-core
protobuf==7.34.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   streamlit
psutil==7.2.2
    # via egovault (pyproject.toml)
pyarrow==23.0.1
    # via streamlit
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycountry==26.2.16
    # via egovault (pyproject.toml)
pycparser==3.0
    # via cffi
pydantic==2.13.0
    # via
    #   mcp
    #   pydantic-settings
pydantic-core==2.46.0
    # via pydantic
pydantic-settings==2.13.1
    # via mcp
pydeck==0.9.1
    # via streamlit
pygments==2.20.0
    # via rich
pyjwt[crypto]==2.12.1
    # via mcp
pyparsing==3.3.2
    # via httplib2
pypdfium2==5.7.0
    # via pdfplumber
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.2.0
    # via egovault (pyproject.toml)
python-dotenv==1.2.2
    # via pydantic-settings
python-multipart==0.0.26
    # via mcp
python-pptx==1.0.2
    # via egovault (pyproject.toml)
pywin32==311
    # via mcp
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.4.4
    # via transformers
requests==2.33.1
    # via
    #   google-api-core
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==15.0.0
    # via
    #   egovault (pyproject.toml)
    #   typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.7.0
    # via transformers
scikit-learn==1.8.0
    # via sentence-transformers
scipy==1.17.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.4.0
    # via egovault (pyproject.toml)
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   ebooklib
    #   python-dateutil
smmap==5.0.3
    # via gitdb
soupsieve==2.8.3
    # via beautifulsoup4
sse-starlette==3.3.4
    # via mcp
starlette==1.0.0
    # via
    #   mcp
    #   sse-starlette
streamlit==1.56.0
    # via egovault (pyproject.toml)
sympy==1.14.0
    # via torch
tenacity==9.1.4
    # via streamlit
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.11.0
    # via sentence-transformers
tornado==6.5.5
    # via streamlit
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==5.5.4
    # via sentence-transformers
typer==0.24.1
    # via
    #   huggingface-hub
    #   transformers
typing-extensions==4.15.0
    # via
    #   altair
    #   beautifulsoup4
    #   huggingface-hub
    #   mcp
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   python-pptx
    #   sentence-transformers
    #   streamlit
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2026.1
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
uvicorn==0.44.0
    # via mcp
watchdog==6.0.0
    # via streamlit
wcwidth==0.6.0
    # via prompt-toolkit
xlsxwriter==3.2.9
    # via python-pptx

# The following packages are considered to be unsafe in a requirements file:
# setuptools
